addpkg(main/kubecolor): Colorize your kubectl output
The github.com/kubecolor/kubecolor repo is the successor to the well-known
github.com/hidetatz/kubecolor project.

Kubecolor wraps kubectl by forwarding the commands to it
and colorizing its output.

Due to kubecolor being written in Go, it requires an Android
build to be able to execute other binaries (in this case kubecolor), due to
Android disallowing certain syscalls (faccessat2) that are used on Linux.